refactor: create FlatpakManifestFinder
Removes some clutter in utils.ts and uses vscode fs api where applicable
