+++ src/utils.ts
@@ -1,10 +1,6 @@
 import * as dbus from 'dbus-next'
 import { promises as fs, constants as fsc } from 'fs'
-import * as path from 'path'
-import { commands, window, workspace, Uri } from 'vscode'
-import * as yaml from 'js-yaml'
-import { FlatpakManifestSchema } from './flatpak.types'
-import { FlatpakManifest } from './flatpakManifest'
+import { commands } from 'vscode'
 
 export const ensureDocumentsPortal = async (): Promise<void> => {
   try {
@@ -13,79 +9,9 @@ export const ensureDocumentsPortal = async (): Promise<void> => {
     const portal = obj.getInterface('org.freedesktop.portal.Documents')
     await portal.GetMountPoint()
   } catch (err) {
-    console.log('Failed to ensure documents portal')
-    console.log(err)
+    console.warn('Failed to ensure documents portal')
+    console.warn(err)
   }
-}
-
-export const isFlatpak = (manifest: FlatpakManifestSchema): boolean => {
-  const hasId = (manifest.id || manifest['app-id']) !== undefined
-  const hasModules = manifest.modules !== undefined
-  return hasId && hasModules
-}
-
-export const parseManifest = async (
-  uri: Uri,
-  isSandboxed: boolean
-): Promise<FlatpakManifest | null> => {
-  const data = (await fs.readFile(uri.fsPath)).toString()
-  let manifest = null
-
-  switch (path.extname(uri.fsPath)) {
-    case '.json':
-      manifest = JSON.parse(data) as FlatpakManifestSchema
-      break
-    case '.yml':
-    case '.yaml':
-      manifest = yaml.safeLoad(data) as FlatpakManifestSchema
-      break
-    default:
-      window
-        .showErrorMessage(
-          'Failed to parse the manifest, please use a valid extension.'
-        ).
-        then(
-          () => { }, // eslint-disable-line @typescript-eslint/no-empty-function
-          () => { } // eslint-disable-line @typescript-eslint/no-empty-function
-        )
-      break
-  }
-
-  if (manifest === null) {
-    return null
-  }
-
-  if (isFlatpak(manifest)) {
-    return new FlatpakManifest(
-      uri,
-      manifest,
-      isSandboxed
-    )
-  }
-
-  return null
-}
-
-export const findManifests = async (
-  isSandboxed: boolean
-): Promise<FlatpakManifest[]> => {
-  const uris: Uri[] = await workspace.findFiles(
-    '**/*.{json,yaml,yml}',
-    '**/{target,.vscode,.flatpak-builder,flatpak_app,.flatpak}/*',
-    1000
-  )
-  const manifests = []
-  for (const uri of uris) {
-    try {
-      const manifest = await parseManifest(uri, isSandboxed)
-      if (manifest) {
-        manifests.push(manifest)
-      }
-    } catch (err) {
-      console.warn(`Failed to parse the manifest at ${uri.fsPath}`)
-    }
-  }
-  return manifests
 }
 
 export const setContext = (ctx: string, state: boolean | string): void => {
